+++ dev-test/gql-tag-operations/gql/index.ts
@@ -1,5 +1,6 @@
 /* eslint-disable */
 import * as graphql from './graphql';
+import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/core';
 
 const documents = {
   '\n  query Foo {\n    Tweets {\n      id\n    }\n  }\n': graphql.FooDocument,
@@ -21,3 +22,10 @@ export function gql(source: string): unknown;
 export function gql(source: string) {
   return (documents as any)[source] ?? {};
 }
+
+export type DocumentType<TDocumentNode extends DocumentNode<any, any>> = TDocumentNode extends DocumentNode<
+  infer TType,
+  any
+>
+  ? TType
+  : never;