+++ packages/graphql-codegen-cli/src/utils/watcher.ts
@@ -118,7 +118,10 @@ export const createWatcher = (
       lifecycleHooks(config.hooks).onWatchTriggered(eventName, path);
       debugLog(`[Watcher] triggered due to a file ${eventName} event: ${path}`);
       const fullPath = join(process.cwd(), path);
-      delete require.cache[fullPath];
+      // In ESM require is not defined
+      try {
+        delete require.cache[fullPath];
+      } catch (err) {}
 
       if (eventName === 'change' && config.configFilePath && fullPath === config.configFilePath) {
         log(`${logSymbols.info} Config file has changed, reloading...`);