+++ packages/presets/gql-tag-operations/tests/fixtures/simple-lowercase-operation-name.ts
@@ -1,0 +1,24 @@
+/* eslint-disable @typescript-eslint/no-unused-vars-experimental, @typescript-eslint/ban-ts-comment */
+//@ts-ignore
+import gql from 'gql';
+
+//@ts-ignore
+const A = gql`
+  query a {
+    a
+  }
+`;
+
+//@ts-ignore
+const B = gql`
+  query b {
+    b
+  }
+`;
+
+//@ts-ignore
+const C = gql`
+  query c {
+    c
+  }
+`;