+++ packages/graphql-cli-codegen-plugin/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @graphql-cli/codegen
 
+## 2.1.1
+
+### Patch Changes
+
+- Updated dependencies [d3c556f8e]
+  - @graphql-codegen/cli@2.1.1
+
+## 2.1.0
+
+### Minor Changes
+
+- 440172cfe: support ESM
+
+### Patch Changes
+
+- Updated dependencies [24185985a]
+- Updated dependencies [39773f59b]
+- Updated dependencies [72044c1bd]
+- Updated dependencies [440172cfe]
+  - @graphql-codegen/cli@2.1.0
+
 ## 2.0.1
 
 ### Patch Changes