+++ packages/plugins/other/schema-ast/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @graphql-codegen/schema-ast
 
+## 2.1.0
+
+### Minor Changes
+
+- 39773f59b: enhance(plugins): use getDocumentNodeFromSchema and other utilities from @graphql-tools/utils
+- 440172cfe: support ESM
+
+### Patch Changes
+
+- 24185985a: bump graphql-tools package versions
+- Updated dependencies [24185985a]
+- Updated dependencies [39773f59b]
+- Updated dependencies [440172cfe]
+  - @graphql-codegen/plugin-helpers@2.1.0
+
 ## 2.0.0
 
 ### Major Changes